+++ src/utils_lgpl/esmfsm/packages/esmfsm_c/src/esm/esm.cpp
@@ -328,7 +328,7 @@ void *
 ESM_Alloc (
     int     contextid,
     char *  name,
-    int     size
+    size_t  size
     ) {
 
     void * return_value = NULL;
